Added pooling of BLS
