+++ occultence/lightcurve/make_lightcurve.py
@@ -510,6 +510,7 @@ class LightCurve:
         was_injected_planet_recovered,
         full_injection_recovery,
         single_injection_recovery,
+        single_clean_detrend,
         split_lightcurve,
         was_planet_observed,
     )
