+++ occultence/recovery/__init__.py
@@ -1,3 +1,4 @@
+from .single_inj_rec import *
 from .injection_recovery import *
 from .stats import *
 from .plots import *
