+++ occultence/version.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.9"
+__version__ = "0.4.0"
 
 
 def version():
@@ -9,3 +9,4 @@ def version():
 # v0.2.2 - added split_time
 # v0.3.0 - added mcmc and lsq
 # v0.3.2 - added celerite and emcee and corner as requirements
+# v0.4.0 - fixed multiprocessing pooling
